+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Send Traffic over ADSL using uplinkTraffic message
 - Low poer mode during development
 - Allow for per protocol RX or TX, or RX and TX selection
+- Show in aircraft tracker the aircraft that it is tracking
+- Adaptive protocol prioritisation. This will allocate more listening slots for protocols that is actually received to increase pings per aircraft 
 
 ### Changed